requirements.txt
pydantic
Flask
outlines[openai]
python-dotenv
openai
mcp[cli]